#17 Move comparables into builtins
